Move sink config bits into separate module
